+++ multivar_horner/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding:utf-8 -*-
 from multivar_horner.classes.abstract_poly import load_pickle
 from multivar_horner.classes.horner_poly import HornerMultivarPolynomial, HornerMultivarPolynomialOpt
 from multivar_horner.classes.regular_poly import MultivarPolynomial